+++ dbmodel/schemas/main/common/base/fct/gw_fct_admin_sys_version_register.sql
@@ -55,7 +55,7 @@ BEGIN
 	END IF;
 
 	IF v_isnew IS NOT TRUE AND v_prev IS NOT NULL THEN
-		v_language := COALESCE(v_language, v_prev.language);
+		v_language := COALESCE(v_prev.language, v_language);
 		v_epsg := COALESCE(v_epsg, v_prev.epsg);
 		v_projecttype := COALESCE(v_projecttype, v_prev.project_type);
 		v_creation_profile := COALESCE(